refactor(llm): :art: LLM service structure refactored.
