--- llmservice/llm-service.js
@@ -1,113 +1,0 @@
-const axios = require('axios');
-const express = require('express');
-
-const app = express();
-const port = 8005;
-
-// Middleware to parse JSON in request body
-app.use(express.json());
-// Load enviornment variables
-require('dotenv').config();
-
-// Define configurations for different LLM APIs
-const llmConfigs = {
-  gemini: {
-    url: (apiKey) => `https://generativelanguage.googleapis.com/v1beta/models/gemini-1.5-flash:generateContent?key=${apiKey}`,
-    transformRequest: (question) => ({
-      contents: [{ parts: [{ text: question }] }]
-    }),
-    transformResponse: (response) => response.data.candidates[0]?.content?.parts[0]?.text
-  },
-  empathy: {
-    url: () => 'https://empathyai.prod.empathy.co/v1/chat/completions',
-    transformRequest: (question) => ({
-      model: "mistralai/Mistral-7B-Instruct-v0.3",
-      messages: [
-        { role: "system", content: "You are a helpful assistant." },
-        { role: "user", content: question }
-      ]
-    }),
-    transformResponse: (response) => response.data.choices[0]?.message?.content,
-    headers: (apiKey) => ({
-      Authorization: `Bearer ${apiKey}`,
-      'Content-Type': 'application/json'
-    })
-  }
-};
-
-// Function to validate required fields in the request body
-function validateRequiredFields(req, requiredFields) {
-  for (const field of requiredFields) {
-    if (!(field in req.body)) {
-      throw new Error(`Missing required field: ${field}`);
-    }
-  }
-}
-
-async function getMusicianName() {
-  try {
-    const response = await axios.get('http://localhost:8004/musicians');
-    const musicianName = response.data.musicianName;
-    console.log('Musician Name:', musicianName);
-    return musicianName;
-  } catch (error) {
-    console.error('Error fetching musician name:', error.message);
-    return null;
-  }
-}
-
-// It does not work as intented. It will be updated
-//getMusicianName();
-// Generic function to send questions to LLM
-async function sendQuestionToLLM(question, apiKey, model = 'gemini') {
-  try {
-    const config = llmConfigs[model];
-    if (!config) {
-      throw new Error(`Model "${model}" is not supported.`);
-    }
-
-    const url = config.url(apiKey);
-    const requestData = config.transformRequest(question);
-
-    const headers = {
-      'Content-Type': 'application/json',
-      ...(config.headers ? config.headers(apiKey) : {})
-    };
-
-    const response = await axios.post(url, requestData, { headers });
-
-    return config.transformResponse(response);
-
-  } catch (error) {
-    console.error(`Error sending question to ${model}:`, error.message || error);
-    return null;
-  }
-}
-
-app.post('/ask', async (req, res) => {
-  try {
-    // Check if required fields are present in the request body
-    validateRequiredFields(req, ['question', 'model']);
-
-    const { question, model } = req.body;
-    //load the api key from an environment variable
-    const apiKey = process.env.LLM_API_KEY;
-    
-    if (!apiKey) {
-      return res.status(400).json({ error: 'API key is missing.' });
-    }
-    const answer = await sendQuestionToLLM(question, apiKey, model);
-    res.json({ answer });
-
-  } catch (error) {
-    res.status(400).json({ error: error.message });
-  }
-});
-
-const server = app.listen(port, () => {
-  console.log(`LLM Service listening at http://localhost:${port}`);
-});
-
-module.exports = server
-
-
