+++ llmservice/package.json
@@ -1,13 +1,14 @@
 {
   "name": "llmservice",
   "version": "1.0.0",
+  "description": " LLM service, in charge of sending queries to the LLMs",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node llm-service.js",
+    "start": "node src/index.js",
     "test": "jest"
   },
   "author": "",
   "license": "ISC",
-  "description": "",
   "homepage": "https://github.com/arquisoft/wichat_en1c#readme",
   "dependencies": {
     "axios": "^1.7.9",
