+++ llmservice/src/config.js
@@ -1,0 +1,4 @@
+// Configuration data for the llm service
+module.exports = {
+    port: process.env.PORT || 8005
+};
