+++ llmservice/test/llm-service.test.js
@@ -3,7 +3,7 @@ process.env.LLM_API_KEY = 'test-api-key';
 
 const request = require('supertest');
 const axios = require('axios');
-const app = require('./llm-service'); 
+const app = require('../src'); 
 
 afterAll(async () => {
     app.close();
